refactor(card_table_panel): split CardTablePanel into mixin package
Move widgets/panels/card_table_panel.py into a package: frame.py keeps
the zone-label constants and UI builders; properties.py holds grid_width
and get_selected_card; handlers.py owns show_loading, set_cards,
_update_panels, click/selection handling, focus/collapse, and image
refresh.
