+++ widgets/panels/card_table_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Card table panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.card_table_panel.frame import CardTablePanel
+
+__all__ = ["CardTablePanel"]
